[api-gateway,database] refactor: Standardize API response types and improve user ID parameter handling
  - Convert user ID parameters from Long to String across all relation endpoints
  - Add new RelationResponse and RelationData DTOs for type-safe API responses
  - Replace generic CommonResponse with specific RelationResponse for relation operations
  - Implement proper error handling for invalid user ID formats with NumberFormatException
  - Update OpenAPI documentation schemas to reflect new response types
  - Enhance service layer to return detailed relation data in responses
  - Maintain backward compatibility while improving API consistency

+++ ripple-database/src/main/java/com/fanaujie/ripple/database/model/UserRelation.java
@@ -10,9 +10,9 @@ import java.time.Instant;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserRelation {
-    public static final byte FRIEND_FLAG = 0x01;
-    public static final byte BLOCKED_FLAG = 0x02;
-    public static final byte HIDDEN_FLAG = 0x04;
+    public static final int FRIEND_FLAG = 0x01;
+    public static final int BLOCKED_FLAG = 0x02;
+    public static final int HIDDEN_FLAG = 0x04;
     
     private long id;
     private long sourceUserId;
